+++ x/apps/types/pool.go
@@ -16,7 +16,7 @@ const (
 type StakingPool Pool
 
 // NewPool creates a new Tokens instance used for queries
-func NewPool(tokens sdk.Int) Pool {
+func NewPool(tokens sdk.BigInt) Pool {
 	return Pool{
 		Tokens: tokens,
 	}
